+++ changelog.d/17600.misc
@@ -1,0 +1,1 @@
+Make the sliding sync `PerConnectionState` class immutable.
